requirements.txt
langchain
ipykernel
python-dotenv
langchain_community
pypdf
beautifulsoup4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
streamlit
langchain_groq
langchain_core
fastapi
uvicorn
langserve
sse-starlette


